pyproject.toml
[tool.poetry]
name = "photoshare"
version = "0.1.0"
description = ""
authors = ["Kirienko-Alyona <ladybird.miami@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
fastapi = "^0.95.1"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
alembic = "^1.10.4"
sqlalchemy = "^2.0.12"
psycopg2 = "^2.9.6"
python-multipart = "^0.0.6"
cloudinary = "^1.32.0"
fastapi-limiter = "^0.1.5"
fastapi-mail = "^1.2.8"
qrcode = "^7.4.2"
pydantic = {extras = ["dotenv"], version = "^1.10.7"}
jinja2 = "^3.1.2"
redis = "^4.5.5"


[tool.poetry.group.dev.dependencies]
pytest-mock = "^3.10.0"
pytest-cov = "^4.0.0"
sphinx = "^7.0.0"


[tool.poetry.group.test.dependencies]
httpx = "^0.24.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]
